feat: #39 number 유틸함수 파일단위로 분리
