+++ src/utils/number/decommaizeNumber.ts
@@ -1,0 +1,3 @@
+export const decommaizeNumber = (numStr: string) => {
+  return Number(numStr.replace(/,/g, ''));
+};
