+++ src/utils/number/index.ts
@@ -1,0 +1,8 @@
+export { commaizeNumber } from './commaizeNumber';
+export { decommaizeNumber } from './decommaizeNumber';
+export { formatPhoneNumber } from './formatPhoneNumber';
+export { isInteger } from './isInteger';
+export { isNegativeInteger } from './isNegativeInteger';
+export { isPositiveInteger } from './isPositiveInteger';
+export { percent } from './percent';
+export { sum } from './sum';
