+++ src/utils/number/isInteger.ts
@@ -1,0 +1,3 @@
+export const isInteger = (num: number) => {
+  return Number.isInteger(num);
+};
