+++ src/utils/number/isNegativeInteger.ts
@@ -1,0 +1,5 @@
+import { isInteger } from './isInteger';
+
+export const isNegativeInteger = (num: number) => {
+  return isInteger(num) && num < 0;
+};
