+++ src/utils/number/isPositiveInteger.ts
@@ -1,0 +1,5 @@
+import { isInteger } from './isInteger';
+
+export const isPositiveInteger = (num: number) => {
+  return isInteger(num) && num > 0;
+};
