+++ src/utils/number/sum.ts
@@ -1,0 +1,3 @@
+export const sum = (nums: number[]) => {
+  return nums.reduce((acc, num) => acc + num, 0);
+};
